Different output if author and committer differ
Sometimes you have commits that come in where the committer is different
than the author. This happens frequently on Open Source projects or if
you happen to have a bot that's automatically committing. It would be
helpful to distinguish commits coming in just from the author and ones
from a committer/bot.